Remove blank book chapter on "How to use salsa"
The chapter was blank. Fwiw, I think it's more
helpful to have only non-blank chapters.

When I first looked at the docs, the message
I got from seeing a blank "how to use" was
"oh, there is no guidance on how to use this thing".
However, there turns out to be plenty of guidance,
with detailled, runnable examples, some API documentation,
and RFCs.

If there is still an aspiration for a high-level
"how to use Salsa" that goes beyond the "Hello World"
example, I'd be happy to adapt that aspiration into a
GitHub issue.