feat: Re-add volatile queries
I have some values which I would like store but which do not implement `Eq`, `volatile` queries, which used to exist, avoid this requirement by simply assuming that the value always changes and therefore do not need to compare it.

An alternative to this which is mentioned in #218 is that we might allow the equality check to be customized via, say `#[salsa::eq(my_eq)]`. It seems a bit more complicated though since the storage assumes it can use `#[derive(Default)]` to construct right now so some codegen changes are needed.

cc #56 #218