Remove unnecessary `Mutex` from singleton initialization
If we ever run into the initialization of a singleton twice, we will panic either way, so there is no need to guard this with a lock.
Instead we can just race in this scenario and then panic when the slower thread tries to set the value of the singleton.