Fully qualify std Result type
The prior macro expansion could produce errors if the macros were
called in a context where `Result` is redefined, for example in a
crate with its own `Result` type which pre-fills the error type. This
replaces existing `Result` uses with `std::result::Result` to avoid
the compilation error in that case.

Signed-off-by: Andrew Lilley Brinker <alilleybrinker@gmail.com>